Gaussian tests working
