+++ .travis.yml
@@ -4,22 +4,7 @@ os:
     - linux
 
 julia:
-    - 1.1
-    - 1.2
-    - nightly
-
-addons:
-  apt:
-    packages:
-      - python-matplotlib
-      - python3-matplotlib
-env:
-  matrix:
-    - PYTHON=python3 # python 3.5
-
-matrix:
-    allow_failures:
-        - julia: nightly
+    - 1.3
 
 fast_finish: true
 notifications:
