+++ Project.toml
@@ -4,28 +4,12 @@ authors = ["Niklas Heim <heim.niklas@gmail.com>"]
 version = "0.1.0"
 
 [deps]
-BSON = "fbb218c0-5317-5bc6-957e-2ee96dd4b1f0"
-CUDAapi = "3895d2a7-ec45-59b8-82bb-cfc6a382f9b3"
+CUDAnative = "be33ccc6-a3ff-5ff2-a52e-74243cff1e17"
 CuArrays = "3a865a2d-5b23-5a0f-bc46-62713ec82fae"
-DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
-DiffEqFlux = "aae7a2af-3d4f-5e19-a356-7da93b79d9d0"
 DrWatson = "634d3b9d-ee7a-5ddf-bec9-22491ea816e1"
 Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
+IRTools = "7869d1d1-7146-5819-86e3-90919afe41df"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-ValueHistories = "98cad3c8-aec3-5f06-8e41-884608649ab7"
-
-[extras]
-CUDAapi = "3895d2a7-ec45-59b8-82bb-cfc6a382f9b3"
-Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
-Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
-Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
-Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Test", "Logging", "Parameters", "Suppressor", "CUDAapi", "Revise"]
+Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
+ZygoteRules = "700de1a5-db45-46bc-99cf-38207098b444"
