+++ src/pdfs/abstract_pdfs.jl
@@ -121,35 +121,4 @@ kld(p::AbstractCPDF, q::AbstractPDF, z::AbstractArray) = error("Not implemented!
 Compute the maximum mean discrepancy between a conditional PDF `p` given `z` 
 and a PDF `q`, given kernel `k`.
 """
-mmd(p::AbstractCPDF, q::AbstractPDF, z::AbstractArray, k) = mmd(k, rand(p,z), rand(q, size(z,2)))    
-
-
-function detect_mapping_variant(x::AbstractVector, xlength::Int) where T
-    if size(x) == (xlength,)
-        return UnitVar
-    elseif size(x) == (xlength+1,)
-        return ScalarVar
-    elseif size(x) == (xlength*2,)
-        return DiagVar
-    else
-        error("Mapping output could not be matched with any variance type.")
-    end
-end
-
-function detect_mapping_variant(x::AbstractMatrix, xlength::Int) where T
-    detect_mapping_variant(x[:,1], xlength)
-end
-
-function detect_mapping_variant(mapping, xlength::Int, zlength::Int)
-    p = first(params(mapping)).data
-    z = randn(zlength, 1)
-    z = isa(p, Array) ? z : z |> gpu
-    x = mapping(z)
-    detect_mapping_variant(x, xlength)
-end
-
-function detect_mapping_variant(mapping::Function, T::Type, xlength::Int, zlength::Int)
-    z = param(randn(T, zlength, 1))
-    x = mapping(z)
-    detect_mapping_variant(x, xlength)
-end
+mmd(p::AbstractCPDF, q::AbstractPDF, z::AbstractArray, k) = mmd(k, rand(p,z), rand(q, size(z,2)))
